+++ executor/revoke.go
@@ -22,6 +22,7 @@ import (
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/mysql"
+	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/util/sqlexec"
 )
@@ -75,7 +76,7 @@ func (e *RevokeExec) Next() (*Row, error) {
 		}
 	}
 	e.done = true
-	notifyUpdatePrivilege(e.ctx)
+	sessionctx.GetDomain(e.ctx).NotifyUpdatePrivilege(e.ctx)
 	return nil, nil
 }
